Added hardware test folder and first test code
First working code to do a hardware test. A test which print ascii character 33 through 126 on the screen. A loop which grows a square on the screen and a loop which makes the whole screen: red , green, blue and white. There is a timing bug in this code the waits are not the same.